HTTP helper methods; break out http helpers and config from the entrypoint file.
